+++ src/main/java/com/karankumar/bookproject/ui/shelf/BooksInShelfView.java
@@ -21,6 +21,9 @@ import com.karankumar.bookproject.backend.service.BookService;
 import com.karankumar.bookproject.backend.service.PredefinedShelfService;
 import com.karankumar.bookproject.ui.MainView;
 import com.karankumar.bookproject.ui.book.BookForm;
+import com.karankumar.bookproject.ui.shelf.listener.BookDeleteListener;
+import com.karankumar.bookproject.ui.shelf.listener.BookGridListener;
+import com.karankumar.bookproject.ui.shelf.listener.BookSaveListener;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.combobox.ComboBox;
 import com.vaadin.flow.component.grid.Grid;
