+++ src/main/java/com/karankumar/bookproject/ui/shelf/listener/BookSaveListener.java
@@ -1,7 +1,8 @@
-package com.karankumar.bookproject.ui.shelf;
+package com.karankumar.bookproject.ui.shelf.listener;
 
 import com.karankumar.bookproject.backend.service.BookService;
 import com.karankumar.bookproject.ui.book.BookForm;
+import com.karankumar.bookproject.ui.shelf.BooksInShelfView;
 import lombok.extern.java.Log;
 
 import java.util.logging.Level;
